+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Authentication against multiple cloud instances also against
   a dynamic database in addition to the static text file
 - Experimental support for saslauthd protocol
+- Credentails caching
+- Connection/request timout option (default: 5s)
 ### Fixed
 ### Changed
 - `external_cloud.*` has been renamed to `xcauth.*` everywhere. :warning: You will also need to rename your configuration file, the old name is deprecated and disappear soon.